third_party/WebKit/LayoutTests/http/tests/devtools/console/console-viewport-stick-to-bottom-expected.txt

Verifies viewport stick-to-bottom behavior.


Running: verifyViewportIsTallEnough
Viewport contains 13 messages

Running: testScrollViewportToBottom
Is at bottom: true, should stick: true

Running: testConsoleSticksToBottom
Is at bottom: true, should stick: true

Running: testEscShouldNotJumpToBottom
Is at bottom: false, should stick: false

Running: testChangingPromptTextShouldJumpToBottom
Is at bottom: true, should stick: true

Running: testViewportMutationsShouldPreserveStickToBottom
Is at bottom: true, should stick: true
Is at bottom: false, should stick: false

Running: testMuteUpdatesWhileScrolling
New messages were muted: true
Refresh was scheduled after dirty state

Running: testShouldNotJumpToBottomWhenPromptFillsEntireViewport
Is at bottom: false, should stick: true
